dmctl: use dmctl binary directly in integration test (#135)
