+++ dm/ctl/master/sql_skip.go
@@ -18,10 +18,11 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/pingcap/dm/dm/ctl/common"
-	"github.com/pingcap/dm/dm/pb"
 	"github.com/pingcap/errors"
 	"github.com/spf13/cobra"
+
+	"github.com/pingcap/dm/dm/ctl/common"
+	"github.com/pingcap/dm/dm/pb"
 )
 
 // NewSQLSkipCmd creates a SQLSkip command
