--- dm/ctl/master/stop_relay.go
@@ -1,59 +1,0 @@
-// Copyright 2019 PingCAP, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package master
-
-import (
-	"fmt"
-
-	"github.com/pingcap/dm/dm/ctl/common"
-	"github.com/pingcap/dm/dm/pb"
-	"github.com/pingcap/errors"
-	"github.com/spf13/cobra"
-)
-
-// NewStopRelayCmd creates a StopRelay command
-func NewStopRelayCmd() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "stop-relay <-w worker ...>",
-		Short: "stop dm-worker's relay unit",
-		Run:   stopRelayFunc,
-	}
-	return cmd
-}
-
-// stopRelayFunc does stop relay request
-func stopRelayFunc(cmd *cobra.Command, _ []string) {
-	if len(cmd.Flags().Args()) > 0 {
-		fmt.Println(cmd.Usage())
-		return
-	}
-
-	workers, err := common.GetWorkerArgs(cmd)
-	if err != nil {
-		common.PrintLines("%s", errors.ErrorStack(err))
-		return
-	}
-	if len(workers) == 0 {
-		fmt.Println("must specify at least one dm-worker (`-w` / `--worker`)")
-		return
-	}
-
-	resp, err := operateRelay(pb.RelayOp_StopRelay, workers)
-	if err != nil {
-		common.PrintLines("can not stop relay unit:\n%v", errors.ErrorStack(err))
-		return
-	}
-
-	common.PrettyPrintResponse(resp)
-}
