+++ tests/dmctl_advance/check_list/sql_inject.sh
@@ -1,0 +1,1 @@
+#!/bin/bash
