+++ tests/dmctl_basic/conf/diff_config.toml
@@ -1,0 +1,57 @@
+# diff Configuration.
+
+log-level = "info"
+
+chunk-size = 10
+
+check-thread-count = 4
+
+sample-percent = 100
+
+use-rowid = false
+
+use-checksum = true
+
+fix-sql-file = "fix.sql"
+
+# tables need to check.
+[[check-tables]]
+schema = "dmctl"
+tables = ["t_target"]
+
+[[table-config]]
+schema = "dmctl"
+table = "t_target"
+remove-columns = ["id"]
+is-sharding = true
+index-fields = "b"
+
+[[table-config.source-tables]]
+instance-id = "source-1"
+schema = "dmctl"
+table  = "~t_.*"
+
+[[table-config.source-tables]]
+instance-id = "source-2"
+schema = "dmctl"
+table  = "~t_.*"
+
+[[source-db]]
+host = "127.0.0.1"
+port = 3306
+user = "root"
+password = ""
+instance-id = "source-1"
+
+[[source-db]]
+host = "127.0.0.1"
+port = 3307
+user = "root"
+password = ""
+instance-id = "source-2"
+
+[target-db]
+host = "127.0.0.1"
+port = 4000
+user = "root"
+password = ""
