+++ tests/dmctl_basic/conf/dm-task.yaml
@@ -1,0 +1,87 @@
+---
+name: test
+task-mode: all
+is-sharding: true
+meta-schema: "dm_meta"
+remove-meta: false
+enable-heartbeat: true
+timezone: "Asia/Shanghai"
+
+target-database:
+  host: "127.0.0.1"
+  port: 4000
+  user: "root"
+  password: ""
+
+mysql-instances:
+  - source-id: "mysql-replica-01"
+    server-id: 101
+    black-white-list:  "instance"
+    route-rules: ["sharding-route-rules-table", "sharding-route-rules-schema"]
+    column-mapping-rules: ["instance-1"]
+    mydumper-config-name: "global"
+    loader-config-name: "global"
+    syncer-config-name: "global"
+
+  - source-id: "mysql-replica-02"
+    server-id: 102
+    black-white-list:  "instance"
+    route-rules: ["sharding-route-rules-table", "sharding-route-rules-schema"]
+    column-mapping-rules: ["instance-2"]
+    mydumper-config-name: "global"
+    loader-config-name: "global"
+    syncer-config-name: "global"
+
+black-white-list:
+  instance:
+    do-dbs: ["dmctl"]
+    do-tables:
+    -  db-name: "dmctl"
+       tbl-name: "~^t_[\\d]+"
+
+routes:
+  sharding-route-rules-table:
+    schema-pattern: dmctl
+    table-pattern: t_*
+    target-schema: dmctl
+    target-table: t_target
+
+  sharding-route-rules-schema:
+    schema-pattern: dmctl
+    target-schema: dmctl
+
+column-mappings:
+  instance-1:
+    schema-pattern: "dmctl"
+    table-pattern: "t_*"
+    expression: "partition id"
+    source-column: "id"
+    target-column: "id"
+    arguments: ["1", "", "t_"]
+
+  instance-2:
+    schema-pattern: "dmctl"
+    table-pattern: "t_*"
+    expression: "partition id"
+    source-column: "id"
+    target-column: "id"
+    arguments: ["2", "", "t_"]
+
+mydumpers:
+  global:
+    mydumper-path: "./bin/mydumper"
+    threads: 4
+    chunk-filesize: 64
+    skip-tz-utc: true
+    extra-args: "--regex '^dmctl.*'"
+
+loaders:
+  global:
+    pool-size: 16
+    dir: "./dumped_data"
+
+syncers:
+  global:
+    worker-count: 16
+    batch: 100
+    max-retry: 100
