+++ tests/dmctl_basic/conf/dm-task2.yaml
@@ -1,0 +1,2 @@
+---
+description: "invalid task config"
