+++ tests/dmctl_basic/conf/dm-worker2.toml
@@ -1,0 +1,15 @@
+# Worker Configuration.
+
+server-id = 102
+source-id = "mysql-replica-02"
+flavor = "mysql"
+meta-file = "relay.meta"
+enable-gtid = false
+relay-binlog-name = ""
+relay-binlog-gtid = ""
+
+[from]
+host = "127.0.0.1"
+user = "root"
+password = ""
+port = 3307
