+++ tests/dmctl_basic/data/db1.increment.sql
@@ -1,0 +1,18 @@
+UPDATE `dmctl`.`t_2` SET `c` = 'lnVaSBHOTnAatvUmHZIZ' WHERE `id` = 6;
+INSERT INTO `dmctl`.`t_1` (`b`,`c`,`d`,`id`) VALUES (1207302373,'slOcgVAMoRUmxBpRdUHe','h',10);
+UPDATE `dmctl`.`t_1` SET `c` = 'uvSLJRmOHYLDDMM' WHERE `id` = 9;
+DELETE FROM `dmctl`.`t_1` WHERE `id` = 2;
+INSERT INTO `dmctl`.`t_1` (`b`,`c`,`d`,`id`) VALUES (617131993,'Klf','dZLD',11);
+UPDATE `dmctl`.`t_2` SET `d` = 'k' WHERE `id` = 3;
+UPDATE `dmctl`.`t_1` SET `d` = 'Nlh' WHERE `id` = 4;
+INSERT INTO `dmctl`.`t_1` (`b`,`c`,`d`,`id`) VALUES (241968486,'LT','JRvYDYFX',12);
+UPDATE `dmctl`.`t_1` SET `c` = 'BZbKSuZKitYG' WHERE `id` = 7;
+DELETE FROM `dmctl`.`t_2` WHERE `id` = 3;
+INSERT INTO `dmctl`.`t_2` (`b`,`c`,`d`,`id`) VALUES (207449096,'upYYtysTjztxkYGprAHD','drBN',11);
+UPDATE `dmctl`.`t_2` SET `c` = 'AVKxHwiUZHXXKqODQoE' WHERE `id` = 8;
+UPDATE `dmctl`.`t_2` SET `d` = 'eLmCZwCA' WHERE `id` = 9;
+INSERT INTO `dmctl`.`t_2` (`b`,`c`,`d`,`id`) VALUES (683005299,'lsATr','mNOpyfu',12);
+UPDATE `dmctl`.`t_2` SET `d` = 'uVNNR' WHERE `id` = 6;
+DELETE FROM `dmctl`.`t_2` WHERE `id` = 5;
+INSERT INTO `dmctl`.`t_1` (`b`,`c`,`d`,`id`) VALUES (373155634,'uoxJTwvCSEuaRbWGq','xzCteQqZwU',13);
+UPDATE `dmctl`.`t_1` SET `d` = 'GzYx' WHERE `id` = 6;
