+++ tests/dmctl_basic/data/db2.increment.sql
@@ -1,0 +1,24 @@
+UPDATE `dmctl`.`t_2` SET `d` = 'S' WHERE `id` = 8;
+DELETE FROM `dmctl`.`t_1` WHERE `id` = 7;
+UPDATE `dmctl`.`t_2` SET `c` = 'pvHeDtXdNQBzc' WHERE `id` = 9;
+INSERT INTO `dmctl`.`t_1` (`b`,`c`,`d`,`id`) VALUES (1533333567,'BEY','nQE',10);
+UPDATE `dmctl`.`t_1` SET `c` = 'UcRjvtdgeAMmmTdd' WHERE `id` = 2;
+DELETE FROM `dmctl`.`t_1` WHERE `id` = 7;
+UPDATE `dmctl`.`t_1` SET `c` = 'oZTluknwkIgUYEkPdB' WHERE `id` = 7;
+UPDATE `dmctl`.`t_1` SET `c` = 'ZcjYaJnh' WHERE `id` = 5;
+DELETE FROM `dmctl`.`t_2` WHERE `id` = 6;
+UPDATE `dmctl`.`t_1` SET `d` = 'QVAzYmqx' WHERE `id` = 4;
+INSERT INTO `dmctl`.`t_1` (`b`,`c`,`d`,`id`) VALUES (815122675,'AkDfRfFpyuMPyU','nOZwdjK',11);
+UPDATE `dmctl`.`t_1` SET `c` = 'SNceJrNMU' WHERE `id` = 6;
+DELETE FROM `dmctl`.`t_2` WHERE `id` = 9;
+UPDATE `dmctl`.`t_1` SET `d` = 'VACozkE' WHERE `id` = 11;
+UPDATE `dmctl`.`t_2` SET `d` = 'Xczvw' WHERE `id` = 4;
+DELETE FROM `dmctl`.`t_1` WHERE `id` = 2;
+UPDATE `dmctl`.`t_2` SET `c` = 'WgFBYecVDGxyytTQNF' WHERE `id` = 7;
+INSERT INTO `dmctl`.`t_1` (`b`,`c`,`d`,`id`) VALUES (1929704764,'hKXJjNu','PzUETJSZdU',12);
+UPDATE `dmctl`.`t_1` SET `d` = 'ZMdgI' WHERE `id` = 4;
+DELETE FROM `dmctl`.`t_2` WHERE `id` = 1;
+UPDATE `dmctl`.`t_1` SET `c` = 'oP' WHERE `id` = 9;
+UPDATE `dmctl`.`t_1` SET `c` = 'fNxtLysT' WHERE `id` = 10;
+DELETE FROM `dmctl`.`t_1` WHERE `id` = 3;
+UPDATE `dmctl`.`t_1` SET `d` = 'gQjjJwwJ' WHERE `id` = 10;
