+++ tests/dmctl_basic/data/db2.increment2.sql
@@ -1,0 +1,18 @@
+alter table `dmctl`.`t_1` drop column d;
+alter table `dmctl`.`t_2` drop column d;
+UPDATE `dmctl`.`t_1` SET `c` = 'ex' WHERE `id` = 12;
+DELETE FROM `dmctl`.`t_1` WHERE `id` = 5;
+UPDATE `dmctl`.`t_2` SET `c` = 'KOhDazkRuLagpqgq' WHERE `id` = 7;
+INSERT INTO `dmctl`.`t_1` (`b`,`c`,`id`) VALUES (2046853891,'YiPcpg',13);
+UPDATE `dmctl`.`t_1` SET `c` = 'TRBVSNLTT' WHERE `id` = 5;
+DELETE FROM `dmctl`.`t_2` WHERE `id` = 3;
+UPDATE `dmctl`.`t_2` SET `c` = 'pEUUHDOepmEqQ' WHERE `id` = 3;
+UPDATE `dmctl`.`t_1` SET `c` = 'yYVYjZd' WHERE `id` = 1;
+DELETE FROM `dmctl`.`t_2` WHERE `id` = 8;
+UPDATE `dmctl`.`t_2` SET `c` = 'kQPauwPFQgoNkBK' WHERE `id` = 2;
+INSERT INTO `dmctl`.`t_2` (`b`,`c`,`id`) VALUES (1328787895,'zGpfwmts',9);
+UPDATE `dmctl`.`t_1` SET `c` = 'CDzG' WHERE `id` = 12;
+DELETE FROM `dmctl`.`t_1` WHERE `id` = 1;
+UPDATE `dmctl`.`t_2` SET `c` = 'IWAyZsgmEdrO' WHERE `id` = 2;
+UPDATE `dmctl`.`t_1` SET `c` = 'OQyXDeydLVAEIwrRrTG' WHERE `id` = 11;
+DELETE FROM `dmctl`.`t_2` WHERE `id` = 2;
