+++ tests/dmctl_basic/data/db2.prepare.sql
@@ -1,0 +1,23 @@
+drop database if exists `dmctl`;
+create database `dmctl`;
+use `dmctl`;
+create table t_1(id bigint auto_increment, b int, c varchar(20), d varchar(10), primary key id(id), unique key b(b));
+create table t_2(id bigint auto_increment, b int, c varchar(20), d varchar(10), primary key id(id), unique key b(b));
+INSERT INTO `dmctl`.`t_2` (`b`,`c`,`d`,`id`) VALUES (536247973,'FdkqoydEmwL','cnmrQHkMMD',1);
+INSERT INTO `dmctl`.`t_1` (`b`,`c`,`d`,`id`) VALUES (625546463,'dgwvtEydzZa','rvuaa',1);
+INSERT INTO `dmctl`.`t_1` (`b`,`c`,`d`,`id`) VALUES (1339059405,'DS','rRbbWfXU',2);
+INSERT INTO `dmctl`.`t_2` (`b`,`c`,`d`,`id`) VALUES (1021801725,'izDZdmGeoeYpyn','YHX',2);
+INSERT INTO `dmctl`.`t_2` (`b`,`c`,`d`,`id`) VALUES (1547191572,'qkdHwHe','OZI',3);
+INSERT INTO `dmctl`.`t_1` (`b`,`c`,`d`,`id`) VALUES (804011834,'cDVHrSBD','ZGzWk',3);
+INSERT INTO `dmctl`.`t_2` (`b`,`c`,`d`,`id`) VALUES (1258473138,'nnTUUdsTOnYsHD','Nz',4);
+INSERT INTO `dmctl`.`t_1` (`b`,`c`,`d`,`id`) VALUES (161128134,'arowzAebwoHBc','Wnux',4);
+INSERT INTO `dmctl`.`t_1` (`b`,`c`,`d`,`id`) VALUES (783636444,'OYlOy','BMkikT',5);
+INSERT INTO `dmctl`.`t_1` (`b`,`c`,`d`,`id`) VALUES (1272018652,'DYpcqArfWolhOhiab','QyzbNqC',6);
+INSERT INTO `dmctl`.`t_1` (`b`,`c`,`d`,`id`) VALUES (1674064434,'HcgEkwijLzTp','CKAuh',7);
+INSERT INTO `dmctl`.`t_1` (`b`,`c`,`d`,`id`) VALUES (275890580,'iroFHGFVsqQyTOvILGTB','jyo',8);
+INSERT INTO `dmctl`.`t_2` (`b`,`c`,`d`,`id`) VALUES (751446388,'MJcpP','aNMJ',5);
+INSERT INTO `dmctl`.`t_1` (`b`,`c`,`d`,`id`) VALUES (844436321,'IxWEwsmadiKId','fw',9);
+INSERT INTO `dmctl`.`t_2` (`b`,`c`,`d`,`id`) VALUES (1394310997,'DLRg','KzhFXaFV',6);
+INSERT INTO `dmctl`.`t_2` (`b`,`c`,`d`,`id`) VALUES (399622170,'sE','DwDwBOEiJ',7);
+INSERT INTO `dmctl`.`t_2` (`b`,`c`,`d`,`id`) VALUES (584262288,'caXN','SN',8);
+INSERT INTO `dmctl`.`t_2` (`b`,`c`,`d`,`id`) VALUES (611154797,'zcsTrw','WRqA',9);
