+++ tests/load_interrupt/run.sh
@@ -49,7 +49,7 @@ function run() {
     run_dm_master $WORK_DIR/master $MASTER_PORT $cur/conf/dm-master.toml
     check_rpc_alive $cur/../bin/check_master_online 127.0.0.1:$MASTER_PORT
 
-    $cur/../bin/dmctl_start_task "$cur/conf/dm-task.yaml"
+    dmctl_start_task
 
     check_port_offline $WORKER1_PORT 20
     check_port_offline $WORKER2_PORT 20
