+++ tests/sharding/run.sh
@@ -30,7 +30,7 @@ function run() {
     check_rpc_alive $cur/../bin/check_master_online 127.0.0.1:$MASTER_PORT
 
     # start DM task only
-    $cur/../bin/dmctl_start_task "$cur/conf/dm-task.yaml"
+    dmctl_start_task
 
     # TODO: check sharding partition id
     # use sync_diff_inspector to check full dump loader
